+++ packages/network/lib/hooks.dart
@@ -3,4 +3,4 @@ export 'src/request.dart';
 export 'src/response.dart';
 export 'src/settings.dart';
 export 'src/hooks.dart';
-export 'src/middleware.dart';
+export 'src/interceptor.dart';